+++ src/js/JSONEditor.js
@@ -5,7 +5,7 @@ const VanillaPicker = require('./vanilla-picker') // may be undefined in case of
 const { treeModeMixins } = require('./treemode')
 const { textModeMixins } = require('./textmode')
 const { previewModeMixins } = require('./previewmode')
-const { clear, extend, getInternetExplorerVersion, parse } = require('./util')
+const { clear, extend, getInnerText, getInternetExplorerVersion, parse } = require('./util')
 const { tryRequireAjv } = require('./tryRequireAjv')
 const { showTransformModal } = require('./showTransformModal')
 const { showSortModal } = require('./showSortModal')
@@ -171,7 +171,7 @@ function JSONEditor (container, options, json) {
  */
 JSONEditor.modes = {}
 
-// debounce interval for JSON schema vaidation in milliseconds
+// debounce interval for JSON schema validation in milliseconds
 JSONEditor.prototype.DEBOUNCE_INTERVAL = 150
 
 JSONEditor.VALID_OPTIONS = [
@@ -184,7 +184,7 @@ JSONEditor.VALID_OPTIONS = [
   'colorPicker', 'onColorPicker',
   'timestampTag', 'timestampFormat',
   'escapeUnicode', 'history', 'search', 'mode', 'modes', 'name', 'indentation',
-  'sortObjectKeys', 'navigationBar', 'statusBar', 'mainMenuBar', 'languages', 'language', 'enableSort', 'enableTransform',
+  'sortObjectKeys', 'navigationBar', 'statusBar', 'mainMenuBar', 'languages', 'language', 'enableSort', 'enableTransform', 'limitDragging',
   'maxVisibleChilds', 'onValidationError',
   'modalAnchor', 'popupAnchor',
   'createQuery', 'executeQuery', 'queryDescription'
@@ -484,6 +484,7 @@ JSONEditor.VanillaPicker = VanillaPicker
 // expose some utils (this is undocumented, unofficial)
 JSONEditor.showTransformModal = showTransformModal
 JSONEditor.showSortModal = showSortModal
+JSONEditor.getInnerText = getInnerText
 
 // default export for TypeScript ES6 projects
 JSONEditor.default = JSONEditor