+++ src/js/previewmode.js
@@ -1,6 +1,6 @@
 'use strict'
 
-import { translate } from './i18n'
+import { setLanguage, setLanguages, translate } from './i18n'
 import { ModeSwitcher } from './ModeSwitcher'
 import { ErrorTable } from './ErrorTable'
 import { showSortModal } from './showSortModal'
@@ -56,6 +56,10 @@ previewmode.create = function (container, options = {}) {
     this.indentation = 2 // number of spaces
   }
 
+  // language
+  setLanguages(this.options.languages)
+  setLanguage(this.options.language)
+
   // determine mode
   this.mode = 'preview'
 
@@ -96,7 +100,7 @@ previewmode.create = function (container, options = {}) {
   this.dom.busy = document.createElement('div')
   this.dom.busy.className = 'jsoneditor-busy'
   this.dom.busyContent = document.createElement('span')
-  this.dom.busyContent.innerHTML = 'busy...'
+  this.dom.busyContent.textContent = 'busy...'
   this.dom.busy.appendChild(this.dom.busyContent)
   this.content.appendChild(this.dom.busy)