+++ test/util.test.js
@@ -96,7 +96,7 @@ describe('util', () => {
       assert.strictEqual(repair('callback({}'), 'callback({}')
     })
 
-    it('should strip trailing zeros', () => {
+    it('should strip trailing commas', () => {
       // matching
       assert.strictEqual(repair('[1,2,3,]'), '[1,2,3]')
       assert.strictEqual(repair('[1,2,3,\n]'), '[1,2,3\n]')
@@ -152,6 +152,34 @@ describe('util', () => {
         '}'
 
       assert.strictEqual(repair(pythonDocument), expectedJson)
+    })
+
+    it('should repair missing comma between objects', () => {
+      const text = '{"aray": [{}{}]}'
+      const expected = '{"aray": [{},{}]}'
+
+      assert.strictEqual(repair(text), expected)
+    })
+
+    it('should not repair normal array with comma separated objects', () => {
+      const text = '[\n{},\n{}\n]'
+
+      assert.strictEqual(repair(text), text)
+    })
+
+    it('should repair line separated json (for example from robo3t', () => {
+      const text = '' +
+        '/* 1 */\n' +
+        '{}\n' +
+        '\n' +
+        '/* 2 */\n' +
+        '{}\n' +
+        '\n' +
+        '/* 3 */\n' +
+        '{}\n'
+      const expected = '[\n{},\n\n{},\n\n{}\n\n]'
+
+      assert.strictEqual(repair(text), expected)
     })
   })